test: move gridsearch tests to their own file, move NestedDict into core
