+++ src/gen_experiments/config.py
@@ -5,13 +5,9 @@ import numpy as np
 import pysindy as ps
 
 from gen_experiments.data import _signal_avg_power
-from gen_experiments.gridsearch.typing import (
-    NestedDict,
-    SeriesDef,
-    SeriesList,
-    SkinnySpecs,
-)
+from gen_experiments.gridsearch.typing import SeriesDef, SeriesList, SkinnySpecs
 from gen_experiments.plotting import _PlotPrefs
+from gen_experiments.typing import NestedDict
 from gen_experiments.utils import FullSINDyTrialData
 
 T = TypeVar("T")
